+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/ReconcilerUtilsTest.java
@@ -6,8 +6,6 @@ package io.strimzi.operator.cluster.operator.assembly;
 
 import io.fabric8.kubernetes.api.model.LabelSelector;
 import io.fabric8.kubernetes.api.model.LabelSelectorBuilder;
-import io.fabric8.kubernetes.api.model.OwnerReference;
-import io.fabric8.kubernetes.api.model.OwnerReferenceBuilder;
 import io.fabric8.kubernetes.api.model.Pod;
 import io.fabric8.kubernetes.api.model.PodBuilder;
 import io.fabric8.kubernetes.api.model.Secret;
@@ -23,6 +21,7 @@ import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticatio
 import io.strimzi.api.kafka.model.common.authentication.KafkaClientAuthenticationScramSha512;
 import io.strimzi.api.kafka.model.kafka.KafkaClusterSpec;
 import io.strimzi.api.kafka.model.kafka.KafkaClusterSpecBuilder;
+import io.strimzi.operator.cluster.ResourceUtils;
 import io.strimzi.operator.cluster.model.NodeRef;
 import io.strimzi.operator.cluster.model.jmx.JmxModel;
 import io.strimzi.operator.cluster.model.jmx.SupportsJmx;
@@ -66,14 +65,6 @@ import static org.mockito.Mockito.when;
 public class ReconcilerUtilsTest {
     private final static String NAME = "my-jmx-secret";
     private final static String NAMESPACE = "namespace";
-    private static final OwnerReference OWNER_REFERENCE = new OwnerReferenceBuilder()
-            .withApiVersion("v1")
-            .withKind("my-kind")
-            .withName("my-name")
-            .withUid("my-uid")
-            .withBlockOwnerDeletion(false)
-            .withController(false)
-            .build();
     private static final Labels LABELS = Labels
             .forStrimziKind("my-kind")
             .withStrimziName("my-name")
@@ -162,7 +153,7 @@ public class ReconcilerUtilsTest {
     @Test
     public void testDisabledJmxWithMissingSecret(VertxTestContext context) {
         KafkaClusterSpec spec = new KafkaClusterSpecBuilder().build();
-        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, OWNER_REFERENCE, spec);
+        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, ResourceUtils.DUMMY_OWNER_REFERENCE, spec);
 
         SecretOperator mockSecretOps = mock(SecretOperator.class);
         when(mockSecretOps.getAsync(eq(NAMESPACE), eq(NAME))).thenReturn(Future.succeededFuture());
@@ -179,7 +170,7 @@ public class ReconcilerUtilsTest {
     @Test
     public void testDisabledJmxWithExistingSecret(VertxTestContext context) {
         KafkaClusterSpec spec = new KafkaClusterSpecBuilder().build();
-        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, OWNER_REFERENCE, spec);
+        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, ResourceUtils.DUMMY_OWNER_REFERENCE, spec);
 
         SecretOperator mockSecretOps = mock(SecretOperator.class);
         when(mockSecretOps.getAsync(eq(NAMESPACE), eq(NAME))).thenReturn(Future.succeededFuture(EXISTING_JMX_SECRET));
@@ -203,7 +194,7 @@ public class ReconcilerUtilsTest {
     @Test
     public void testEnabledJmxWithoutAuthWithMissingSecret(VertxTestContext context) {
         KafkaClusterSpec spec = new KafkaClusterSpecBuilder().withNewJmxOptions().endJmxOptions().build();
-        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, OWNER_REFERENCE, spec);
+        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, ResourceUtils.DUMMY_OWNER_REFERENCE, spec);
 
         SecretOperator mockSecretOps = mock(SecretOperator.class);
         when(mockSecretOps.getAsync(eq(NAMESPACE), eq(NAME))).thenReturn(Future.succeededFuture());
@@ -220,7 +211,7 @@ public class ReconcilerUtilsTest {
     @Test
     public void testEnabledJmxWithoutAuthWithExistingSecret(VertxTestContext context) {
         KafkaClusterSpec spec = new KafkaClusterSpecBuilder().withNewJmxOptions().endJmxOptions().build();
-        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, OWNER_REFERENCE, spec);
+        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, ResourceUtils.DUMMY_OWNER_REFERENCE, spec);
 
         SecretOperator mockSecretOps = mock(SecretOperator.class);
         when(mockSecretOps.getAsync(eq(NAMESPACE), eq(NAME))).thenReturn(Future.succeededFuture(EXISTING_JMX_SECRET));
@@ -249,7 +240,7 @@ public class ReconcilerUtilsTest {
                     .endKafkaJmxAuthenticationPassword()
                 .endJmxOptions()
                 .build();
-        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, OWNER_REFERENCE, spec);
+        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, ResourceUtils.DUMMY_OWNER_REFERENCE, spec);
 
         SecretOperator mockSecretOps = mock(SecretOperator.class);
         when(mockSecretOps.getAsync(eq(NAMESPACE), eq(NAME))).thenReturn(Future.succeededFuture());
@@ -265,7 +256,7 @@ public class ReconcilerUtilsTest {
                     assertThat(secret, is(notNullValue()));
                     assertThat(secret.getMetadata().getName(), is(NAME));
                     assertThat(secret.getMetadata().getNamespace(), is(NAMESPACE));
-                    assertThat(secret.getMetadata().getOwnerReferences(), is(List.of(OWNER_REFERENCE)));
+                    assertThat(secret.getMetadata().getOwnerReferences(), is(List.of(ResourceUtils.DUMMY_OWNER_REFERENCE)));
                     assertThat(secret.getMetadata().getLabels(), is(LABELS.toMap()));
                     assertThat(secret.getMetadata().getAnnotations(), is(nullValue()));
                     assertThat(secret.getData().size(), is(2));
@@ -284,7 +275,7 @@ public class ReconcilerUtilsTest {
                     .endKafkaJmxAuthenticationPassword()
                 .endJmxOptions()
                 .build();
-        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, OWNER_REFERENCE, spec);
+        JmxModel jmx = new JmxModel(NAMESPACE, NAME, LABELS, ResourceUtils.DUMMY_OWNER_REFERENCE, spec);
 
         SecretOperator mockSecretOps = mock(SecretOperator.class);
         when(mockSecretOps.getAsync(eq(NAMESPACE), eq(NAME))).thenReturn(Future.succeededFuture(EXISTING_JMX_SECRET));
@@ -300,7 +291,7 @@ public class ReconcilerUtilsTest {
                     assertThat(secret, is(notNullValue()));
                     assertThat(secret.getMetadata().getName(), is(NAME));
                     assertThat(secret.getMetadata().getNamespace(), is(NAMESPACE));
-                    assertThat(secret.getMetadata().getOwnerReferences(), is(List.of(OWNER_REFERENCE)));
+                    assertThat(secret.getMetadata().getOwnerReferences(), is(List.of(ResourceUtils.DUMMY_OWNER_REFERENCE)));
                     assertThat(secret.getMetadata().getLabels(), is(LABELS.toMap()));
                     assertThat(secret.getMetadata().getAnnotations(), is(nullValue()));
                     assertThat(secret.getData().size(), is(2));